convert stickygroup tests to vitest
